NormalizingFlows

Documentation for NormalizingFlows.

NormalizingFlows.grad!

NormalizingFlows.loglikelihood

NormalizingFlows.optimize

NormalizingFlows.train_flow

NormalizingFlows.value_and_gradient!

NormalizingFlows.grad! — Method

grad!(
    rng::AbstractRNG,
    ad::ADTypes.AbstractADType,
    vo,
    θ_flat::AbstractVector{<:Real},
    reconstruct,
    out::DiffResults.MutableDiffResult,
    args...
)

Compute the value and gradient for negation of the variational objective vo at θ_flat using the automatic differentiation backend ad.

Default implementation is provided for ad where ad is one of AutoZygote, AutoForwardDiff, AutoReverseDiff (with no compiled tape), and AutoEnzyme. The result is stored in out.

Arguments

rng::AbstractRNG: random number generator

ad::ADTypes.AbstractADType: automatic differentiation backend

vo: variational objective

θ_flat::AbstractVector{<:Real}: flattened parameters of the normalizing flow

reconstruct: function that reconstructs the normalizing flow from the flattened parameters

out::DiffResults.MutableDiffResult: mutable diff result to store the value and gradient

args...: additional arguments for vo

NormalizingFlows.loglikelihood — Method

loglikelihood(flow::Bijectors.TransformedDistribution, xs::AbstractVecOrMat)

Compute the log-likelihood for variational distribution flow at a batch of samples xs from the target distribution.

NormalizingFlows.optimize — Method

optimize(
    rng::AbstractRNG, 
    ad::ADTypes.AbstractADType, 
    vo, 
    θ₀::AbstractVector{T}, 
    re, 
    args...; 
    kwargs...
)

Iteratively updating the parameters θ of the normalizing flow re(θ) by calling grad! and using the given optimiser to compute the steps.

Arguments

rng::AbstractRNG: random number generator

ad::ADTypes.AbstractADType: automatic differentiation backend

vo: variational objective

θ₀::AbstractVector{T}: initial parameters of the normalizing flow

re: function that reconstructs the normalizing flow from the flattened parameters

args...: additional arguments for vo

Keyword Arguments

max_iters::Int=10000: maximum number of iterations

optimiser::Optimisers.AbstractRule=Optimisers.ADAM(): optimiser to compute the steps

show_progress::Bool=true: whether to show the progress bar. The default information printed in the progress bar is the iteration number, the loss value, and the gradient norm.

callback=nothing: callback function with signature cb(iter, opt_state, re, θ) which returns a dictionary-like object of statistics to be displayed in the progress bar. re and θ are used for reconstructing the normalizing flow in case that user want to further axamine the status of the flow.

prog=ProgressMeter.Progress( max_iters; desc="Training", barlen=31, showspeed=true, enabled=show_progress ): progress bar configuration

Returns

θ: trained parameters of the normalizing flow

opt_stats: statistics of the optimiser

st: optimiser state for potential continuation of training

NormalizingFlows.train_flow — Method

train_flow([rng::AbstractRNG, ]vo, flow, args...; kwargs...)

Train the given normalizing flow flow by calling optimize.

Arguments

rng::AbstractRNG: random number generator

vo: variational objective

flow: normalizing flow to be trained

args...: additional arguments for vo

Keyword Arguments

max_iters::Int=1000: maximum number of iterations

optimiser::Optimisers.AbstractRule=Optimisers.ADAM(): optimiser to compute the steps

ADbackend::ADTypes.AbstractADType=ADTypes.AutoZygote(): automatic differentiation backend

kwargs...: additional keyword arguments for optimize (See optimize)

Returns

flow_trained: trained normalizing flow

opt_stats: statistics of the optimiser during the training process (See optimize)

st: optimiser state for potential continuation of training

NormalizingFlows.value_and_gradient! — Function

value_and_gradient!(
    ad::ADTypes.AbstractADType,
    f,
    θ::AbstractVector{T},
    out::DiffResults.MutableDiffResult
) where {T<:Real}

Compute the value and gradient of a function f at θ using the automatic differentiation backend ad. The result is stored in out. The function f must return a scalar value. The gradient is stored in out as a vector of the same length as θ.