Documentation

Documentation

Functions

Index

Functions

NormalizingFlows.AffineLayer — Type

Affine Layer with parameters:

W: Weight of the conditioner

b: Bias of the conditioner

AffineLayer implements the conditioner and the transformation to the input z

NormalizingFlows.GLOW — Type

GLOW has 3 layers: First - Conv1x1 layer (for ease of finding inverse : W = PLU, where P is a fixed permutation layer, L and U are learnt) Second - Normalization Third - Affine Coupling layer

NormalizingFlows.PlanarFlow — Type

The parameters of Planar Flows are as follows:

v ∈ ℝᴰ

w ∈ ℝᴰ

b ∈ ℝ

The transformation is - z′ = z + vσ(wᵀz + b) where σ is an activation function such as tanh, z is the input to the flow and z′ is the output.

This flow can be interpreted as expanding/contracting the space in the direction perpendicular to the hyperplane wᵀz + b = 0

NormalizingFlows.RadialFlow — Type

The parameters of radial flow are:

α α ∈ (0, +∞)

β ∈ ℝ

z₀ ∈ ℝᴰ

It takes the following form: z′ = z + β/(α + r(z))*(z - z₀) where r(z) = ||z-z₀||

The above transformation can be thought of as a contraction/expansion radially with center z₀

NormalizingFlows.affinecouplinglayer — Method

Coupling layer: It transforms the input as:

zᵢ′ = zᵢ for i < d

zᵢ′ = A(zᵢ) for i >= d where d = D/2

Inputs -

z : input from trainig data

A : AffineLayer of the required size d, i.e., D/2 where D is the size of input data

NormalizingFlows.expected_pdf — Method

`expected_pdf(data, p_u, A)`

Inputs

z : Value whose probability density is estimated

pᵤ : Base distribution which may be from the package Distributions

or any distribution which can be sampled using rand

A : Affine layer

Returns the probability density of z wrt to the distribution given by A

NormalizingFlows.loss_kl — Method

KL Divergence

NormalizingFlows.lower_ones — Method

Returns lower triangular square matrix of size k whose values are 1

NormalizingFlows.sample — Method

`sample(pᵤ, A)`

Inputs -

pᵤ : Base distribution which may be from the package Distributions

or any distribution which can be sampled using rand

A : Affine layer

NormalizingFlows.sample — Method

`sample(pᵤ, L)`

Inputs -

pᵤ : Base distribution which may be from the package Distributions

or any distribution which can be sampled using rand

L : Linear Flow

NormalizingFlows.sample — Method

`sample(pᵤ, P)`

Inputs -

pᵤ : Base distribution which may be from the package Distributions or any distribution which can be sampled using rand

P : Planar flow

NormalizingFlows.sample — Method

`sample(pᵤ, R)`

Inputs -

pᵤ : Base distribution which may be from the package Distributions or any distribution which can be sampled using rand

R : Radial flow

NormalizingFlows.train! — Method

`train!(rng::AbstractRNG, data, loss, p_u, opt, model)`

The function updates ps, the parameters of the model, so that pᵤ may be transformed to the distribution of each of the labelled inputs.

Inputs -

rng

data : The training data, sequence of samples of the target distribution

loss : A function which takes the model, input, x and u, sample of base distribution and returns the divergence between the model applied to x and base distribution, which must be minimized.

pᵤ : Base distribution which may be from the package Distributions or any distribution which can be sampled using rand

opt : Flux Optimizer

model : The neural network used, i.e., one among AffineLayer, GLOW, PlanarFlow and RadialFlow

NormalizingFlows.train_labelled_data! — Method

`train_labelled_data!(rng::AbstractRNG, data_loader, pᵤ, opt, model)`

Returns trained models for each category in the labels. The trained models tranform data from the base distribution into the target distribution.

Inputs -

rng

data_loader : The labelled training data, sequence of samples of the target distribution which are labelledwith integers

loss : A function which takes the model, input, x and u, sample of base distribution and returns the divergence between the model applied to x and base distribution, which must be minimized.

pᵤ : Base distribution which may be from the package Distributions or any distribution which can be sampled using rand

opt : Flux Optimizer

model : The neural network used, i.e., one among AffineLayer, GLOW, PlanarFlow and RadialFlow

epochs: Number of times train! runs

Index

NormalizingFlows.AffineLayer

NormalizingFlows.GLOW

NormalizingFlows.PlanarFlow

NormalizingFlows.RadialFlow

NormalizingFlows.affinecouplinglayer

NormalizingFlows.expected_pdf

NormalizingFlows.loss_kl

NormalizingFlows.lower_ones

NormalizingFlows.sample

NormalizingFlows.sample

NormalizingFlows.sample

NormalizingFlows.sample

NormalizingFlows.train!

NormalizingFlows.train_labelled_data!